adding actuary types: context | provider | i18n. Adapting cards to new types
