+++ packages/core/src/types/common/i18n/index.ts
@@ -1,2 +1,3 @@
 export * from './employer';
 export * from './trustee';
+export * from './actuary';
