+++ packages/core/src/types/common/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './employer';
 export * from './trustee';
+export * from './actuary';
