+++ packages/layout/src/components/cards/actuary/actuaryMachine.tsx
@@ -1,5 +1,5 @@
 import { Machine, assign } from 'xstate';
-import { Actuary } from './context';
+import { ActuaryContext } from '@tpr/core';
 
 interface ActuaryStates {
 	states: {
@@ -32,12 +32,12 @@ type ActuaryEvents =
 	| { type: 'BACK' }
 	| { type: 'DELETE' };
 
-export interface ActuaryContext {
-	complete: boolean;
-	remove: { confirm: boolean; date: string } | null;
-	actuary: Partial<Actuary>;
-	preValidatedData?: boolean | null;
-}
+// export interface ActuaryContext {
+// 	complete: boolean;
+// 	remove: { confirm: boolean; date: string } | null;
+// 	actuary: Partial<Actuary>;
+// 	preValidatedData?: boolean | null;
+// }
 
 const actuaryMachine = Machine<ActuaryContext, ActuaryStates, ActuaryEvents>({
 	id: 'actuary',
