+++ packages/layout/src/components/cards/actuary/context.tsx
@@ -1,17 +1,12 @@
-import React, { createContext, useContext, ReactElement } from 'react';
+import React, { createContext, useContext } from 'react';
 import { useMachine } from '@xstate/react';
-import actuaryMachine, { ActuaryContext as AC } from './actuaryMachine';
-import { State, EventData } from 'xstate';
-import { i18n as i18nDefaults, ActuaryI18nProps } from './i18n';
+import actuaryMachine from './actuaryMachine';
+import { i18n as i18nDefaults } from './i18n';
 import { useI18n } from '../hooks/use-i18n';
 import {
-	RecursivePartial,
-	CardDefaultProps,
-	CardPersonalDetails,
-	CardContactDetails,
-	CardAddress,
-	CardProviderProps,
-} from '../common/interfaces';
+	ActuaryContextProps,
+	ActuaryProviderProps
+} from '@tpr/core';
 
 export const ActuaryContext = createContext<ActuaryContextProps>({
 	current: {},
@@ -22,30 +17,6 @@ export const ActuaryContext = createContext<ActuaryContextProps>({
 	onSaveName: Promise.resolve,
 	i18n: i18nDefaults,
 });
-
-type RenderProps = (_props: ActuaryContextProps) => ReactElement;
-
-export interface ActuaryContextProps
-	extends Omit<ActuaryProviderProps, 'actuary'> {
-	send: (event: any, payload?: EventData) => Partial<State<AC, any, any, any>>;
-	current: Partial<State<AC, any, any, any>>;
-}
-
-export interface Actuary
-	extends CardDefaultProps,
-		CardPersonalDetails,
-		CardContactDetails {
-	organisationName: string;
-	address: Partial<CardAddress>;
-}
-
-export interface ActuaryProviderProps extends CardProviderProps {
-	/** Actuary props from the API */
-	actuary: Partial<Actuary>;
-	children?: RenderProps | ReactElement;
-	/** overwrite any text that you need */
-	i18n?: RecursivePartial<ActuaryI18nProps>;
-}
 
 export const ActuaryProvider = ({
 	complete,
