+++ packages/layout/src/components/cards/actuary/i18n.ts
@@ -1,78 +1,4 @@
-export type ActuaryI18nProps = {
-	preview: {
-		buttons: {
-			one: string;
-			two: string;
-			three: string;
-			four: string;
-		};
-		checkboxLabel: string;
-	};
-	name: {
-		title: string;
-		fields: {
-			title: {
-				label: string;
-			};
-			firstName: {
-				label: string;
-				error: string;
-			};
-			lastName: {
-				label: string;
-				error: string;
-			};
-		};
-	};
-	contacts: {
-		title: string;
-		subtitle: string;
-		fields: {
-			telephone: {
-				label: string;
-				error: string;
-			};
-			email: {
-				label: string;
-				error: string;
-			};
-		};
-	};
-	remove: {
-		confirm: {
-			breadcrumbs: {
-				link1: string;
-				link2: string;
-			};
-			title: string;
-			dialog: {
-				message1: string;
-			};
-			buttons: {
-				remove: string;
-				cancel: string;
-			};
-		};
-		date: {
-			title: string;
-			fields: {
-				confirm: {
-					label: string;
-				};
-				date: {
-					label: string;
-					hint: string;
-					error: string;
-				};
-			};
-			errors: {
-				formIncomplete: string;
-				dateAddedBeforeEffectiveDate: string;
-				dateAddedInTheFuture: string;
-			};
-		};
-	};
-};
+import { ActuaryI18nProps } from '@tpr/core';
 
 export const i18n: ActuaryI18nProps = {
 	preview: {
