+++ packages/layout/src/components/cards/actuary/views/name/index.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import { FieldProps } from '@tpr/forms';
 import { useActuaryContext } from '../../context';
-import { ActuaryI18nProps } from '../../i18n';
 import {
 	RecursivePartial,
 	cardType,
 	cardTypeName,
-} from '../../../common/interfaces';
+	ActuaryI18nProps
+} from '@tpr/core';
 import NameForm from '../../../common/views/nameForm/nameForm';
 
 const getFields = (
